Fix attempt trying to clean shoryuken.rb
Still kinda just moving stuff around, but I hope to be able to use
`Shoryuken::Options` as an instance, not as a class
